refactor: created db folder
